pyproject.toml
[tool.poetry]
name = "fastapi-ecommerce"
version = "0.1.0"
description = ""
authors = ["Andrey Ivanchenko <varioustemp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.16.1"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.1.0"
greenlet = "^3.2.2"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
